# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorBody'
      success:
        type: boolean
    type: object
  request.CreateArticleRequest:
    properties:
      category_id:
        type: integer
      content:
        minLength: 1
        type: string
      cover_image:
        maxLength: 500
        type: string
      slug:
        maxLength: 200
        minLength: 1
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      summary:
        maxLength: 500
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
...
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  request.UpdateArticleRequest:
    properties:
      category_id:
        type: integer
      content:
        minLength: 1
        type: string
      cover_image:
        maxLength: 500
        type: string
      slug:
        maxLength: 200
        minLength: 1
        type: string
      status:
        enum:
        - draft
        - published
        type: string
      summary:
        maxLength: 500
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
      version:
        minimum: 1
        type: integer
    required:
    - version
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API 支持
    url: http://www.swagger.io/support
  description: 博客后端 API 服务，支持文章管理、配置文件认证等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 博客系统 API
  version: "1.0"
paths:
  /articles:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 按状态筛选
        in: query
        name: status
        type: string
      - description: 按分类slug筛选
        in: query
        name: category
        type: string
      - description: 按标签slug筛选
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取文章列表
      tags:
      - article
    post:
      consumes:
      - application/json
      parameters:
      - description: 文章数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 创建文章
      tags:
      - article
  /articles/{id}:
    delete:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: 无内容
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 删除文章
      tags:
      - article
    get:
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据 ID 获取文章
      tags:
      - article
    put:
      consumes:
      - application/json
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 更新文章
      tags:
      - article
...
  /articles/random:
    get:
      parameters:
      - default: 5
        description: 数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取随机文章
      tags:
      - article
  /articles/recent:
    get:
      parameters:
      - default: 5
        description: 数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取最近文章
      tags:
      - article
  /articles/search:
    get:
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 搜索文章
      tags:
      - article
  /articles/slug/{slug}:
    get:
      parameters:
      - description: 文章 Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 根据 Slug 获取文章
      tags:
      - article
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录凭据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 用户登录
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: 要撤销的刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 用户登出
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 刷新访问令牌
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证令牌，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"